+++ src/django_errors/apps.py
@@ -7,3 +7,8 @@ class DjangoErrorsConfig(AppConfig):
     """django-errors apps config"""
 
     name = "django_errors"
+
+    def ready(self) -> None:
+        from django_errors import py314_compat
+
+        py314_compat.apply()